se realiza cambios en mapa V2.0, se realizo cambio de look and fill en pantalla de login
