+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
 
         <meta-data
             android:name="com.google.android.geo.API_KEY"
-            android:value="@string/google_maps_key" />
+            android:value="AIzaSyA7d5VxLP4VRayPEO1IAZP9fveTXy44J_A" />
 
         <activity
             android:name=".mapa.RecorridoConductor"
