+++ app/src/main/res/layout/fragment_empleado_camino.xml
@@ -24,22 +24,12 @@
         android:text="Mostrar Ruta"
         app:layout_constraintBottom_toTopOf="@+id/cv_Instrucciones"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.241"
+        app:layout_constraintHorizontal_bias="0.498"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.978"
+        app:layout_constraintVertical_bias="0.976"
         tools:layout_conversion_absoluteHeight="48dp"
         tools:layout_conversion_absoluteWidth="138dp" />
-
-    <Button
-        android:id="@+id/btnAUX"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="12dp"
-        android:text="Button"
-        app:layout_constraintBottom_toTopOf="@+id/cv_Instrucciones"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/btn_Inicio" />
 
     <androidx.cardview.widget.CardView
         android:id="@+id/cv_Instrucciones"
